+++ rllib/algorithms/ppo/torch/ppo_torch_rl_module.py
@@ -1,8 +1,9 @@
+from dataclasses import dataclass
 from typing import Mapping, Any, Union
 
 import gymnasium as gym
 
-from ray.rllib.core.rl_module.rl_module import RLModule
+from ray.rllib.core.rl_module.rl_module import RLModule, RLModuleConfig
 from ray.rllib.core.rl_module.torch import TorchRLModule
 from ray.rllib.models.experimental.base import STATE_OUT
 from ray.rllib.models.experimental.configs import FCConfig, FCEncoderConfig
@@ -20,10 +21,9 @@ from ray.rllib.models.torch.torch_distributions import (
     TorchDiagGaussian,
 )
 from ray.rllib.policy.sample_batch import SampleBatch
-from ray.rllib.utils.annotations import override
+from ray.rllib.utils.annotations import override, ExperimentalAPI
 from ray.rllib.utils.framework import try_import_torch
 from ray.rllib.utils.gym import convert_old_gym_space_to_gymnasium_space
-from ray.rllib.algorithms.ppo.ppo import PPOModuleConfig
 from ray.rllib.utils.nested_dict import NestedDict
 
 
@@ -41,6 +41,30 @@ def get_ppo_loss(fwd_in, fwd_out):
     loss = actor_loss + critic_loss
 
     return loss
+
+
+@ExperimentalAPI
+@dataclass
+class PPOModuleConfig(RLModuleConfig):
+    """Configuration for the PPO RLModule.
+
+    Attributes:
+        observation_space: The observation space of the environment.
+        action_space: The action space of the environment.
+        shared_encoder_config: The configuration for the encoder network.
+        pi_config: The configuration for the policy network.
+        vf_config: The configuration for the value network.
+        free_log_std: For DiagGaussian action distributions, make the second half of
+            the model outputs floating bias variables instead of state-dependent. This
+            only has an effect is using the default fully connected net.
+    """
+
+    observation_space: gym.Space = None
+    action_space: gym.Space = None
+    shared_encoder_config: FCConfig = None
+    pi_config: FCConfig = None
+    vf_config: FCConfig = None
+    free_log_std: bool = False
 
 
 class PPOTorchRLModule(TorchRLModule):
